+++ .release-it.ts
@@ -3,8 +3,10 @@ import type { Config } from 'release-it';
 // See: https://github.com/release-it/release-it/blob/main/docs/recipes/monorepo.md
 
 export default {
-  // @ts-expect-error don't publish root package
-  npm: false,
+  npm: {
+    // don't publish root package
+    publish: false,
+  },
   git: {
     // allow dirty dir, because packages will have version bumped
     requireCleanWorkingDir: false,
@@ -21,7 +23,7 @@ export default {
     },
   },
   hooks: {
-    'before:init': [
+    'before:version:bump': [
       'npm ci',
       'npm run lint',
       'npm run prettier',
@@ -29,7 +31,7 @@ export default {
       'npm run build',
       'npm test',
       // publish all packages, then make repo-related release steps (git tag, GitHub release)
-      'npx turbo run release --ui=tui',
+      'npx turbo run release -- -i ${version} --ci',
     ],
   },
 } satisfies Config;