+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * chore: use relative path in sig
 * fix: check TTL within single test run
 * feat: split values per `runId`
+* feat: add `globalCache.teardownReporter`
 
 ## [0.2.2] - 2025-09-03
 * feat: better signature checks.